Move policy match stuff to its own file
